pyproject.toml
[project]
name = "customer-account-smartsheet"
version = "0.1.0"
description = "Add a short description here"
dependencies = ["pandas~=2.1.3", "smartsheet-python-sdk~=3.0.2", "python-dotenv~=1.0.0"]
readme = "README.md"
requires-python = ">= 3.8"
license = { text = "MIT" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
